Add scala.js support
