+++ .travis.yml
@@ -1,7 +1,12 @@
+sudo: false
 language: scala
 scala:
-  - 2.10.4
-  - 2.11.4
+  - 2.10.5
+  - 2.11.7
+# Temporary to fix travis-ci/travis-ci#4527
+before_script:
+- mkdir -p $HOME/.sbt/launchers/0.13.8/
+- curl -L -o $HOME/.sbt/launchers/0.13.8/sbt-launch.jar http://dl.bintray.com/typesafe/ivy-releases/org.scala-sbt/sbt-launch/0.13.8/sbt-launch.jar
 notifications:
   irc:
     channels:
