+++ build.sbt
@@ -1,6 +1,6 @@
-import com.typesafe.sbt.pgp.PgpKeys.publishSigned
 import sbtrelease.Utilities._
 import sbtunidoc.Plugin.UnidocKeys._
+import ReleaseTransformations._
 
 lazy val buildSettings = Seq(
   organization := "org.spire-math",
@@ -24,14 +24,22 @@ lazy val commonSettings = Seq(
     "-Ywarn-numeric-widen",
     //"-Ywarn-value-discard", // fails with @sp on Unit
     "-Xfuture"
-  )
+  ) ++ (CrossVersion.partialVersion(scalaVersion.value) match {
+    case Some((2, 11)) => Seq("-Ywarn-unused-import")
+    case _             => Seq.empty
+  }),
+  scalacOptions in (Compile, console) ~= (_ filterNot (_ == "-Ywarn-unused-import")),
+  scalacOptions in (Test, console) := (scalacOptions in (Compile, console)).value,
+  scalaJSStage in Test := FastOptStage,
+  fork := false,
+  parallelExecution in Test := false
 )
 
 lazy val algebraSettings = buildSettings ++ commonSettings ++ publishSettings
 
 lazy val docSettings = Seq(
   autoAPIMappings := true,
-  unidocProjectFilter in (ScalaUnidoc, unidoc) := inProjects(core, std, laws),
+  unidocProjectFilter in (ScalaUnidoc, unidoc) := inProjects(coreJVM, stdJVM, lawsJVM),
   site.addMappingsToSiteDir(mappings in (ScalaUnidoc, packageDoc), "api"),
   git.remoteRepo := "git@github.com:non/algebra.git"
 )
@@ -43,33 +51,56 @@ lazy val aggregate = project.in(file("."))
   .settings(site.settings: _*)
   .settings(ghpages.settings: _*)
   .settings(docSettings: _*)
-  .aggregate(core, laws, std)
-  .dependsOn(core, laws, std)
+  .aggregate(coreJVM, lawsJVM, stdJVM)
+  .dependsOn(coreJVM, lawsJVM, stdJVM)
+  .aggregate(coreJS, lawsJS, stdJS)
+  .dependsOn(coreJS, lawsJS, stdJS)
 
-lazy val core = project
+lazy val core = crossProject.crossType(CrossType.Pure)
   .settings(moduleName := "algebra")
   .settings(algebraSettings: _*)
 
-lazy val std = project.dependsOn(core)
+lazy val coreJVM = core.jvm 
+lazy val coreJS = core.js
+
+lazy val std = crossProject
+  .dependsOn(core)
   .settings(moduleName := "algebra-std")
   .settings(algebraSettings: _*)
 
-lazy val laws = project.dependsOn(core, std)
+lazy val stdJVM = std.jvm 
+lazy val stdJS = std.js
+
+lazy val laws = crossProject
+  .dependsOn(core, std, macros)
   .settings(moduleName := "algebra-laws")
   .settings(algebraSettings: _*)
   .settings(
     libraryDependencies ++= Seq(
-      "org.scalacheck" %% "scalacheck" % "1.12.2",
-      "org.typelevel" %% "discipline" % "0.2.1",
-      "org.scalatest" %% "scalatest" % "2.2.4" % "test"
+      "org.scalacheck" %%% "scalacheck" % "1.12.4",
+      "org.typelevel" %%% "discipline" % "0.4",
+      "org.scalatest" %%% "scalatest" % "3.0.0-M7" % "test"
     )
   )
+
+lazy val lawsJVM = laws.jvm 
+lazy val lawsJS = laws.js
+
+lazy val macros = crossProject.crossType(CrossType.Pure)
+  .settings(moduleName := "algebra-macros")
+  .settings(algebraSettings: _*)
+  .settings(crossVersionSharedSources:_*)
+  .settings(scalaMacroDependencies:_*)
+
+lazy val macrosJVM = macros.jvm 
+lazy val macrosJS = macros.js
 
 lazy val publishSettings = Seq(
   homepage := Some(url("http://spire-math.org")),
   licenses := Seq("MIT" -> url("http://opensource.org/licenses/MIT")),
   autoAPIMappings := true,
   apiURL := Some(url("https://non.github.io/algebra/api/")),
+
   releaseCrossBuild := true,
   releasePublishArtifactsAction := PgpKeys.publishSigned.value,
   publishMavenStyle := true,
@@ -77,6 +108,7 @@ lazy val publishSettings = Seq(
   pomIncludeRepository := Function.const(false),
   publishTo <<= (version).apply { v =>
     val nexus = "https://oss.sonatype.org/"
+
     if (v.trim.endsWith("SNAPSHOT"))
       Some("snapshots" at nexus + "content/repositories/snapshots")
     else
@@ -109,8 +141,25 @@ lazy val publishSettings = Seq(
         <url>http://github.com/tixxit/</url>
       </developer>
     </developers>
+  ),
+  releaseProcess := Seq[ReleaseStep](
+    checkSnapshotDependencies,
+    inquireVersions,
+    runTest,
+    setReleaseVersion,
+    commitReleaseVersion,
+    tagRelease,
+    publishArtifacts,
+    setNextVersion,
+    commitNextVersion,
+    ReleaseStep(action = Command.process("sonatypeReleaseAll", _)),
+    pushChanges
   )
 )
+
+addCommandAlias("validate", ";compile;test;unidoc")
+
+// Base Build Settings
 
 lazy val noPublishSettings = Seq(
   publish := (),
@@ -118,4 +167,35 @@ lazy val noPublishSettings = Seq(
   publishArtifact := false
 )
 
-addCommandAlias("validate", ";compile;test;unidoc")
+def crossVersionSharedSources() =
+  Seq(Compile, Test).map { sc =>
+    (unmanagedSourceDirectories in sc) ++= {
+      (unmanagedSourceDirectories in sc ).value.map {
+        dir:File => new File(dir.getPath + "_" + scalaBinaryVersion.value)
+      }
+    }
+  }
+
+lazy val scalaMacroDependencies: Seq[Setting[_]] = Seq(
+  libraryDependencies += "org.scala-lang" % "scala-reflect" % scalaVersion.value % "provided",
+  libraryDependencies ++= {
+    CrossVersion.partialVersion(scalaVersion.value) match {
+      // if scala 2.11+ is used, quasiquotes are merged into scala-reflect
+      case Some((2, scalaMajor)) if scalaMajor >= 11 => Seq()
+      // in Scala 2.10, quasiquotes are provided by macro paradise
+      case Some((2, 10)) =>
+        Seq(
+          compilerPlugin("org.scalamacros" % "paradise" % "2.1.0-M5" cross CrossVersion.full),
+              "org.scalamacros" %% "quasiquotes" % "2.1.0-M5" cross CrossVersion.binary
+        )
+    }
+  }
+)
+
+addCommandAlias("gitSnapshots", ";set version in ThisBuild := git.gitDescribedVersion.value.get + \"-SNAPSHOT\"")
+
+// For Travis CI - see http://www.cakesolutions.net/teamblogs/publishing-artefacts-to-oss-sonatype-nexus-using-sbt-and-travis-ci
+credentials ++= (for {
+  username <- Option(System.getenv().get("SONATYPE_USERNAME"))
+  password <- Option(System.getenv().get("SONATYPE_PASSWORD"))
+} yield Credentials("Sonatype Nexus Repository Manager", "oss.sonatype.org", username, password)).toSeq
