+++ laws/jvm/src/test/scala/algebra/laws/LawTestsJvm.scala
@@ -1,0 +1,13 @@
+package algebra
+package laws
+
+import algebra.macros._
+import algebra.std.Rat
+import CheckSupport._
+
+class LawTests extends LawTestsBase {
+
+  // Rational tests do not return oin Scala-js, so we make them JVM only.
+  laws[BaseLaws, Rat].check(_.isReal)
+  laws[RingLaws, Rat].check(_.field)
+}
