+++ laws/shared/src/main/scala/algebra/laws/BaseLaws.scala
