+++ laws/shared/src/main/scala/algebra/laws/CheckSupport.scala
