+++ laws/shared/src/main/scala/algebra/laws/GroupLaws.scala
