+++ laws/shared/src/main/scala/algebra/laws/LatticeLaws.scala
