+++ laws/shared/src/main/scala/algebra/laws/LatticePartialOrderLaws.scala
