+++ laws/shared/src/main/scala/algebra/laws/LogicLaws.scala
