+++ laws/shared/src/main/scala/algebra/laws/OrderLaws.scala
