+++ laws/shared/src/main/scala/algebra/laws/RingLaws.scala
