+++ laws/shared/src/main/scala/algebra/laws/package.scala
