+++ project/plugins.sbt
@@ -1,5 +1,7 @@
-addSbtPlugin("com.eed3si9n"      % "sbt-unidoc"             % "0.3.2")
-addSbtPlugin("com.github.gseitz" % "sbt-release"            % "1.0.0")
-addSbtPlugin("com.jsuereth"      % "sbt-pgp"                % "1.0.0")
-addSbtPlugin("com.typesafe.sbt"  % "sbt-ghpages"            % "0.5.3")
-addSbtPlugin("com.typesafe.sbt"  % "sbt-site"               % "0.8.1")
+addSbtPlugin("com.eed3si9n"      % "sbt-unidoc"   % "0.3.2")
+addSbtPlugin("com.github.gseitz" % "sbt-release"  % "1.0.0")
+addSbtPlugin("com.jsuereth"      % "sbt-pgp"      % "1.0.0")
+addSbtPlugin("org.xerial.sbt"    % "sbt-sonatype" % "0.5.0")
+addSbtPlugin("com.typesafe.sbt"  % "sbt-ghpages"  % "0.5.3")
+addSbtPlugin("com.typesafe.sbt"  % "sbt-site"     % "0.8.1")
+addSbtPlugin("org.scala-js"      % "sbt-scalajs"  % "0.6.4")
