+++ std/shared/src/main/scala/algebra/std/Util.scala
