+++ std/shared/src/main/scala/algebra/std/all.scala
