+++ std/shared/src/main/scala/algebra/std/boolean.scala
