+++ std/shared/src/main/scala/algebra/std/byte.scala
