+++ std/shared/src/main/scala/algebra/std/char.scala
