+++ std/shared/src/main/scala/algebra/std/int.scala
