+++ std/shared/src/main/scala/algebra/std/map.scala
