+++ std/shared/src/main/scala/algebra/std/rat.scala
