+++ std/shared/src/main/scala/algebra/std/set.scala
