+++ std/shared/src/main/scala/algebra/std/short.scala
