+++ std/shared/src/main/scala/algebra/std/string.scala
